rekolte-api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
PyJWT==2.8.0
google-auth==2.29.0
requests==2.31.0
scikit-learn==1.6.1
xgboost==2.0.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.2
reportlab==4.2.0
gunicorn==22.0.0
